# backend/internal/delivery/http/openapi.yaml
openapi: 3.0.3
info:
  title: Platform RT API
  description: |
    OpenAPI 3.0 specification for the Sitransparan RT/RW backend.

    Autentikasi: `Authorization: Bearer <jwt>` (HS256, valid 24 jam).
    Role: `superadmin`, `admin_rt`, `resident`.
    Konvensi akses pada deskripsi endpoint:
    - PUBLIC: tanpa autentikasi
    - AUTH: semua role terautentikasi
    - ADMIN: superadmin + admin_rt
    - SUPERADMIN: superadmin saja

    Tenant context hanya diambil dari JWT claims yang terverifikasi (header
    X-Tenant-ID, query param, dan subdomain TIDAK dipercaya). Endpoint publik
    `/api/v1/t/{slug}/...` me-resolve tenant dari slug di path.
  version: 2.0.0
servers:
  - url: http://localhost:8081
    description: Local server (Port 8081)
  - url: /
    description: Current host relative path
paths:
...
  /api/v1/users:
    get:
      summary: List user - superadmin scope global, admin_rt scope tenant-nya (ADMIN)
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: limit, in: query, schema: { type: integer } }
        - { name: offset, in: query, schema: { type: integer } }
      responses:
        '200':
          description: '{ data: UserWithRole[], total, limit, offset }'
    post:
      summary: Buat user - hanya superadmin yang boleh membuat role superadmin (ADMIN)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, role]
              properties:
                name: { type: string }
                email: { type: string }
                password: { type: string }
                phone: { type: string }
                role: { type: string, enum: [superadmin, admin_rt, resident] }
                tenant_id: { type: string, format: uuid, description: Hanya dipakai superadmin untuk scope global }
      responses:
        '201':
          description: User dibuat
        '403':
          description: Role escalation ditolak
  /api/v1/users/{id}:
    get:
      summary: Detail user (ADMIN)
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: UserWithRole
    put:
      summary: Update user - role superadmin hanya oleh superadmin (ADMIN)
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, role]
              properties:
                name: { type: string }
                email: { type: string }
                phone: { type: string }
                role: { type: string, enum: [superadmin, admin_rt, resident] }
                password: { type: string }
                tenant_id: { type: string, format: uuid }
      responses:
        '200':
          description: User diupdate
        '403':
          description: Role escalation ditolak
    delete:
      summary: Hapus user (ADMIN)
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: User dihapus

  # ------------------------------------------------------------- Residents
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...